+++ NEWS.md
@@ -7,6 +7,7 @@
 * Add warning message to PlotPerformanceByAttribute() if multiple column names are present in both subass and attributes
 * Fix group colors in PlotPerformanceByAttribute()
 * Update allowable "types" in CompareFiles() to fix bug preventing comparison of certain HYPE file types
+* Update WriteHarmonizedData() and WriteHarmonizedSpatialDescription() to prevent error caused when writing file without loading data.table package
 
 ## HYPEtools 1.6.2 (2024-05-24)
 *Enhancements and bug fixes*